+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- updated dependencies (FastAPI, Langchain, OpenAI)
+
+### Added
+
+- add endpoint `POST /embeddings/generateFromFile` for embedding generation
+- add support for _Azure OpenAI_ provider for embedding generation and LLM usage
+
 ## 0.3.1 - 2024-09-05
 
 ## 0.3.0 - 2024-09-05